insights-processor/requirements.txt
pyspark==3.5.1
pymongo==4.6.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.0